+++ Renommer.Designer.cs
@@ -1,0 +1,89 @@
+﻿namespace Gestionnaire_de_Fond_d_Écran
+{
+    partial class Renommer
+    {
+        /// <summary>
+        /// Required designer variable.
+        /// </summary>
+        private System.ComponentModel.IContainer components = null;
+
+        /// <summary>
+        /// Clean up any resources being used.
+        /// </summary>
+        /// <param name="disposing">true if managed resources should be disposed; otherwise, false.</param>
+        protected override void Dispose(bool disposing)
+        {
+            if (disposing && (components != null))
+            {
+                components.Dispose();
+            }
+            base.Dispose(disposing);
+        }
+
+        #region Windows Form Designer generated code
+
+        /// <summary>
+        /// Required method for Designer support - do not modify
+        /// the contents of this method with the code editor.
+        /// </summary>
+        private void InitializeComponent()
+        {
+            this.btn_valider = new System.Windows.Forms.Button();
+            this.lbl_extension = new System.Windows.Forms.Label();
+            this.txt_nom = new System.Windows.Forms.TextBox();
+            this.SuspendLayout();
+            // 
+            // btn_valider
+            // 
+            this.btn_valider.Location = new System.Drawing.Point(198, 42);
+            this.btn_valider.Name = "btn_valider";
+            this.btn_valider.Size = new System.Drawing.Size(75, 23);
+            this.btn_valider.TabIndex = 0;
+            this.btn_valider.Text = "Valider";
+            this.btn_valider.UseVisualStyleBackColor = true;
+            this.btn_valider.Click += new System.EventHandler(this.btn_valider_Click);
+            // 
+            // lbl_extension
+            // 
+            this.lbl_extension.AutoSize = true;
+            this.lbl_extension.Location = new System.Drawing.Point(238, 16);
+            this.lbl_extension.Name = "lbl_extension";
+            this.lbl_extension.Size = new System.Drawing.Size(24, 13);
+            this.lbl_extension.TabIndex = 1;
+            this.lbl_extension.Text = ".ext";
+            // 
+            // txt_nom
+            // 
+            this.txt_nom.Location = new System.Drawing.Point(13, 13);
+            this.txt_nom.Name = "txt_nom";
+            this.txt_nom.Size = new System.Drawing.Size(222, 20);
+            this.txt_nom.TabIndex = 2;
+            // 
+            // Renommer
+            // 
+            this.AutoScaleDimensions = new System.Drawing.SizeF(6F, 13F);
+            this.AutoScaleMode = System.Windows.Forms.AutoScaleMode.Font;
+            this.ClientSize = new System.Drawing.Size(284, 71);
+            this.Controls.Add(this.txt_nom);
+            this.Controls.Add(this.lbl_extension);
+            this.Controls.Add(this.btn_valider);
+            this.FormBorderStyle = System.Windows.Forms.FormBorderStyle.FixedDialog;
+            this.MaximizeBox = false;
+            this.MinimizeBox = false;
+            this.Name = "Renommer";
+            this.ShowIcon = false;
+            this.ShowInTaskbar = false;
+            this.StartPosition = System.Windows.Forms.FormStartPosition.CenterParent;
+            this.Text = "Renommer...";
+            this.ResumeLayout(false);
+            this.PerformLayout();
+
+        }
+
+        #endregion
+
+        private System.Windows.Forms.Button btn_valider;
+        private System.Windows.Forms.Label lbl_extension;
+        private System.Windows.Forms.TextBox txt_nom;
+    }
+}